fcoeadm: Fix memory leak when scanning for the fc_host
One dirent is not being freed if we find the fc_host or
if our buffer is too small. There's also no bounds checking
as the 'if (n > 0)' is outside the while loop.

The fix is to switch to using a for-statement and to free
the dirent for all directory entries in the array.

Signed-off-by: Robert Love <robert.w.love@intel.com>