fcoe-utils: Don't overwrite /etc/fcoe/config if it exists
If there's already an /etc/fcoe/config file then we shouldn't
overwrite it on a new 'make install'. It's a particular problem
when debugging becuase if you set DEBUG='yes' in config a
make install will overwrite it with DEBUG='no'.

To do this a post install hook is added to the Makefile.am.

Signed-off-by: Robert Love <robert.w.love@intel.com>