fcoeadm: Fix interface validation
We're checking for all sorts of things (fcoemon connection,
sysfs mount) every time we need to validate an interface.
We also have a validation routine that just checks to see
if the string is empty.

This patch makes it so that we're checking overall
requirements (fcoemon, sysfs) in main and then simply
checking if the interface is valid if one was passed in.

This patch splits the Ethernet adapter checking from
the FCoE connection checking. This Ethernet adapter checking
for the 'c' option and both Ethernet adapter and FCoE
conneciton checking for all other options.

A side effect of this patch is that fcoeadm will fail if
fcoemon isn't running even for the information commands
(-i, -t, -l). There are two arguments for this behavior,

1) The fcoe-utils suite requires fcoemon to be running for
   all FCoE connections. If fcoemon isn't running there
   shouldn't be any active connections.

2) It clearly shows the user that there is a service
   failure.

If users start complaining about this behaviour we can revert
it, but I don't see a demand now and at this point it's better
to simplify the userspace functionality until it's more
stable.

Lastly, we don't need to do interface name validation in
fcoeadm_display.c. The getopt loop should be validating
input and then the display code can trust it.

Signed-off-by: Robert Love <robert.w.love@intel.com>