fcoemon: Don't parse cfg-ethx
An example per-interface configuration file is created
on install, the file name is cfg-ethx. Currenly fcoemon
treats "ethx" as a valid interface name. We should skip
this file when we're parsing through the configuration
files.

The only bad behavior that this is causing without this
patch is that we will write "ethx" to
/sys/module/fcoe/parameters/create when stopping the
fcoe service.

Signed-off-by: Robert Love <robert.w.love@intel.com>