fcoeadm: Fix fcoeadm_display.c:display_adapter_info()
This logic is just wrong. If devices are created in the
wrong order then fcoeadm will display incorrect ports under
the wrong adapters.

This patch changes the algorithm for each display_* routine
to be the following-

1) Initialize HBA table
   - Open HBA adapters
   - Read HBA and port info into table

2) Display information
   - Uses already open adapter handle(s)

3) Destroy HBA table
   - Releases adapter handles

The HBA table is no longer global and is instead allocated
by each routine using it (only one allocation per fcoeadm call).

Reorgainizing this way allowed for code consolidation in each
of the display functions where the table of HBAs needed to
be initialized.

Signed-off-by: Robert Love <robert.w.love@intel.com>