fcoeadm: Remove 'struct opt_info'
This structure isn't needed. Passing arguments to display_*
routines makes the code more readable.

The only thing the display_* options need are the ifname
and display_target_info() is used for both LUN and target
displaying so it needs a flag. An enum was created for the
second case. The interface name can just be a pointer that's
assigned to the appropriate argv element since that memory
will not be free'd durring fcoeadm's execution. This makes
things nicer too becuase now we don't need to check string
sizes to see if an ifname was provided. Now we can just check
if the pointer is NULL or not.

Signed-off-by: Robert Love <robert.w.love@intel.com>