fcoeadm: Rename fcoeadm.h to fcoeadm_display.h
This file exposes routines and other symbols necessary
to use the fcoeadm_display routines. It is not a file
that allows other code to interact with fcoeadm.c.

The rename makes it clear that it is the display interface.

This patch also changes some of the includes so that each
.c file only includes what it needs.

Signed-off-by: Robert Love <robert.w.love@intel.com>