fipvlan: convert logging functions to sa_log routines, remove log.[ch]
fipvlan was carrying around it's own debug logging functions.  I'm trying
to prepare for sharing code between fipvlan and fcoemon, which requires the
use of the same error and debug logging functions.

This makes fipvlan link against librt, because of the use of
clock_gettime in fcoemon_utils.c.  That will be fixed.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>