foce-utils: make a real utilities library for shared routines
Once upon a time there were some libraries as part of this source tree.
Most of that code was removed, as it was providing a layer of OS abstraction
that we don't care about.  But in the cleanup process, everything that was
shared was lumped into a couple of object files to be linked into any
executable that needed them.  That has some ugly side effects, like making
any code that wants to share some simple logging functions link in librt,
because the same object file also includes timer functionality.  If the
functionality were split up into multiple object files, which could then be
linked into a library, only what was needed would be linked in.

Bring back a local convenience library for shared routines with limited
dependencies, so I can start doing better code sharing between tools.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>