fipvlan: more robust timeout values when waiting for IFF_RUNNING
It can take a while for a network interface to establish link and negotiate
DCB settings, all of which must be done before FIP VLAN discovery will
complete.  The existing timeout is OK when waiting for a response, but far
too short for waiting for a link to become usable.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>