fcnsq: fix error reporting when the ioctl fails for gpn_id and gnn_id
I ran into this because I had a user/kernel ABI mismatch in the FC BSG
interface[1].  The fix was just to recompile with up-to-date kernel
headers, but I did notice that the error messages were wrong because
these compares are screwed up.

Signed-off-by: Chris Leech <christopher.leech@intel.com>

[1] There was a change to the kernel in Jan that added a packed
attribute to struct fc_bsg_request.  It would have been the right thing
to do in the first place, but it did change ABI 2 kernel releases after
the FC BSG interface was introduced.
Signed-off-by: Robert Love <robert.w.love@intel.com>