fcoe-utils: documentation update
Update the content of the fcoeadm and fcoemon man pages, and add a fipvlan
man page.  This patch now puts the man pages in asciidoc format, with a
Makefile that generates nroff man pages using asciidoc and the DocBook XSL
stylesheets.  The generated nroff pages are also checked into source
control, as users shouldn't need to have the asciidoc and DocBook tools
installed to generate them.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>