fcoe-utils: fix VLAN discovery when the interface is also part of a bond
VLAN discovery would fail if the Ethernet interface was also part of a
bond.  The VLAN discovery code in both fcoemon and fipvlan needed to set
the PACKET_ORIGDEV socket option.  Also, skip over bonding interfaces in
fipvlan so that it doesn't try and run over a bond when using the auto
setting.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>