fcoeadm: remove check for running fcoemon
It's nice to be able to run the informational commands that don't talk to
fcoemon, even if fcoemon isn't running.  For instance, checking to see
which connects were left active after stopping fcoemon.  The commands that
need fcoemon will still fail if it's not running because there will not be
a listening client interface socket.

Plus, this just checks for the existence of a pid file.  It doesn't even
make sure there's a running process with that pid.  If fcoemon doesn't exit
cleanly, fcoeadm will still pass these checks.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>