fcoe-utils: use SAN MAC if available for FIP VLAN Request in lib/fip
Per FC-BB-5, all FIP frames should use the same ENode MAC execpt the
keep-alive message. When the LLD has support for SAN MAC, the FCoE
kernel stack would use it, but all FIPs that happen in user space
currently don't have the knowledge for that.

This adds support to query the SAN MAC and use it for as the FIP
frame source MAC as well as the FIP MAC descriptor.

Signed-off-by: Yi Zou <yi.zou@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>