fcoemon: close fip socket and cancel timers when stopping VLAN discovery
VLAN discovery is restarted after a network interface is removed and the
re-added (network driver reload); but if the FIP socket is reused from
before it will be bound to the wrong ifindex and received frames will
not be correctly routed to fcoemon.  Make sure we destroy the FIP socket
when leaving the VLAN discovery state, so a new one will be created
later when needed.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>