lib/fip.c: make fip_recv() non-blocking
When testing networking driver reloading with fcoemon running, I
occasionally see fcoemon hang in a call to fip_recv.  I don't know why
select is indicating receive data ready when it's not, but this fixes
thing up and keeps fcoemon running.

fip_recv is called in fcoemon and fipvlan, but in both cases only after
a call to poll or select has indicated receive data was ready and
neither should ever need to block.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>