fcoemon: don't clear DCB query setting on VLAN link up
DCB status query isn't restarted on a VLAN link up, because there will also be
a link event for the real device which is used instead.  But currently the
state is set back to INIT, stopping any previous DCB query from running.  This
can cause fcoemon to not create an fcoe instance on a ready VLAN interface
after a link event.  Instead, just don't do anything for VLANs.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>