fipvlan: ensure that a socket is never in the poll list more than once
If the link state is changing while fipvlan is running, it could result in a
packet socket being added to the poll list more than once.  When that happened,
recvmsg would be called multiple times for a single FIP response.

Previously this caused a hang, which the change to MSG_DONTWAIT worked around,
but an error message is still printed every time the condition occurs.  This
change fixes the poll list managment to correctly avoid the error.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>