fipvlan: make sure recv_loop always uses an up-to-date polling list
The poll file descriptor list is updated when a link event is delivered via
a netlink packet, which happens inside of recv_loop.  When the list is
updated, the global pfd and pfd_len values can both change, but recv_loop
continues on using the original values which have been passed in on the
stack.  Instead this patch makes recv_loop reference the global symbols
directly, so that updated values can be used on subsequent calls to poll
from inside the loop.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>