fcoemon: rewind the config file pointer, so that order no longer matters
The config file parsing has a strict order requirement, because the file is
only scanned through once.  Instead, rewind the file before scanning for each
config variable, so that changing the order doesn't break anything.

Add debug prints of the configuration variables are they are read, to make
testing this easier.

Signed-off-by: Chris Leech <christopher.leech@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>