fcoe-utils: Use enum fcoe_status for all status codes
There are multiple enums used for error and status
codes. This causes problems when converting error codes
from fcoemon to fcoeadm. This patch makes it so that
only one enum is used for error and status codes.

checkpatch.pl returns a few WARNINGS due to
non-negative error codes. This should be fine.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>