fcoe-utils: Active auto_vlan=0 ports when interface operstate goes up
When ifconfig down is performed on a vlan interface fcoe-utils calls
DISABLE on the fcoe port. However, when it comes back up, it does not
call ENABLE since it was an auto created port. It waits for vlan
discovery to complete. But since vlan discovery is disabled for vlan
interfaces, ENABLE never gets called and the link remains down.

Note: The current fcoe driver registers a device notifier and calls
fcoe_ctlr_link_up on its own. This patch makes ENABLE/DISABLE symmetric
and avoids forcing the llds to register their own notifiers.

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>