fcoeutils: Retry vlan discovery forever if AUTO_VLAN and FCOE_ENABLE are set
This is a modified patch as per Robert Love's comments -
https://lists.open-fcoe.org/pipermail/devel/2011-March/011272.html

After trying vlan discovery 10 times fcoemon disables auto_vlan on the
port and falls back to the physical interface. bnx2fc does not support
this. This patch checks for FCOE_ENABLE and AUTO_VLAN and if both are
set for the interface, the vlan discovery will continue until a vlan is
discovered.

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>