fcoemon: Unexpected FLOGI after fcoe port is destroyed
When fcoeadm -d is performed, fcoemon sets fcoe_enable=0 on the
physical port being destroyed but not on the vlan port. If a link toggle
occurs at this time, two link events are received, one each for physical
and vlan interfaces. The physical interface event is ignored due to
fcoe_enable=0 but the vlan event is processed and CREATE is called
resulting in unexpected FLOGI. This also puts the physical and vlan
ports out of sync and a subsequent fcoeadm -d is not processed.

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>