fcoe-utils: Drain fip socket during creation to discard unintended rx packets.
During fcoe start, fcoemon creates a PF_PACKET,ETH_P_FIP raw socket for
vlan discovery.  After the socket() call and before binding to the
desired device interface, the socket receives FIP packets from all
interfaces. This can interfere with VLAN discovery if a discovery packet
on one interface is received on the wrong socket. This patch drains the
socket after binding it to an interface and before adding the fd to
select monitoring.

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>