fcoe-utils: Add BASH completion files for fcoeadm and fcoemon
These completion scripts will do general completion of
fcoeadm and fcoemon options. The only real shortcoming is
that fcoeadm cannot complete the <interval> value of the
'--stats' and '-s' options.

These scripts may or may not be added to the bash-completion
project and added to the distro's bash-completion packages.
'make install' will copy these scripts to the
/etc/bash_completion.d/ directory. However, they are not
usable unless the bash-completion package is installed.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>