fcoeadm: Remove mentions to error strings
Having error strings in the man page doesn't necessarily
help the user. Knowing what error string to expect doesn't
allow the user to prevent erronenous behavior. The error
strings should describe the problem well enough that the
user should know what to do.

Additionally, having the error strings in the man page
requires developers to keep the man page in sync with any
fcoeadm string changes, which is a pain.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>