fcoe-utils: modify fcoemon to use lldpad abstract socket
lldpad has been modified to use an abstract namespace socket
for its client interface.  This patch modifies fcoemon to
use this new socket type to connect to lldpad.

NOTE: The change requires that the file clif_sock.h
from the lldpad package has the corresponding changes to
support the abstract namespace socket - in which:
LLDP_CLIF_SOCK is defined as "/com/intel/lldpad"

From: Hannes Reinecke <hare@suse.de>

For abstract sockets we need to calculate the address length
correctly so as not to be confused with trailing characters.
And we don't need to unlink anything here.

[ Robert Love updated this patch with a lldpad version
  dependency check in configure.ac and some documentation
  updates to ensure that this package is build with a
  version of lldpad compatible with this change. ]

Signed-off-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>