fcoeadm: Validate interface for '-S' and '-r' options
These options require arguments. This patch validates that
the arguments are both valid interfaces and that there is a
fcoe connection on them before proceeding to issue a command
to fcoemon.

This changes the error message to the user. Previously the
command would be sent to fcoemon, which would fail and fcoeadm
would return "Internal Error." Now that validation is occuring
and fcoemon is not involved we return a more appropriate, "No
connection created on <ifname>" message.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>