fcoe-utils: Avoid internal error on fcoeadm -d
When a fcoeadm -d operation is performed on an interface that has no
port created, fcoemon times out waiting for response. This is because
since the last_action is set to FCP_DESTROY_IF, the state machine sets
the next action to FCP_WAIT and does not send a response back. To avoid
this if the last_action is FCP_DESTROY_IF, we return an explicit
ENOACTION.

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>