fcoe-utils: Maintain ready state of vlan port for processing link updates
Currently there is no relation maintained between a physical interface
and its dependent vlan interfaces. Due to this the vlan fcoe port can go
out of sync with the physical fcoe port and result in undesirable state
machine behaviour. E.g. handling link events when it shouldn't. A ready
flag is introduced that is set to false when the physical port is
destroyed/disabled. Link events are not processed for the vlan ports
when not ready.

v2: Fix missing brace for the auto_created condition check

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>