fcoeadm: Increase clif timeout to a reasonably large value
During testing we routinely hit "internal error" messages when working
with a number of ports = 1024 and when the kernel is busy. Although the
operation completes successfully, the message is confusing to the user.
Increase the timeout to 30s to allow normal workloads to finish without
error. This is a select timeout and will not cause long waits if the
command completes early since select will return.

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Yi Zou <yi.zou@intel.com>