fipvlan: fix auto mode for bonding device
The fipvlan discovery can be sent out to the master bond but will come back on
the active slave, where the slave interface as tracked by fipvlan interfaces
list still thinks it's not IFF_RUNNING and no FIP VLAN DISC is sent. Rather,
we will not populate the bonding master to the list, but allow FIP VLAN DISC
to be sent over all slaves. Also correctly check if the FIP VLAN DISC is
already sent on an interface to avoid repeated sending the same frame. Another
check is added via lookup_fcf() to avoid populate already discovered FCFs.

Signed-off-by: Yi Zou <yi.zou@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>