fcoemon: fix the logging of fc host netlink event
The currently logging for scsi netlink message for fc host event is not helpful
at all. Add proper logging of the event type and its translations that match the
kernel scsi_transport_fc, so the log can actually be interpreted and used for
debugging. Move ARRAY_SIZE into the common fcoemon header so all can use and
remove repeated defines of it in several other places.

Signed-off-by: Yi Zou <yi.zou@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>