fcoemon: Double free of sock_reply pointer when create fails
Calling create twice on the same port results in the 2nd create failing
with a "connection already exists" message. However the fcm_cli_create
function still sets the sock_reply pointer to reply sockinfo structure.
On return to the caller, the reply pointer is freed because the create
returns an error code. The invalid sock_reply pointer is then again
freed in fcm_handle_changes resulting in corruption.

The fix is to set sock_reply only on a successful create.

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>