fcoemon: fix fcoemon create/destroy a fipvlan auto created interface
For fcoe interfaces created by 'fipvlan -cs', it is not supposed to be managed
by fcoeadm/fcoemon. However, the bug in fcoemon currently allow this to
happend by doing:
1). fcoeadm -c ethx
2). fcoeadm -d ethx

1) would fail the create but still add the fcoe_port to the fcoe port list
managed by fcoemon due to not checking if there is a already existing
interface. 2) would succeed as the interface (created via fipvlan not fcoeadm)
is not managed by fcoemon after 1).

Fix this by double-checking if we already have an fchost associated with the
input interface before allocating a new fcoe port and issuing the create
command to the sysfs.

Signed-off-by: Yi Zou <yi.zou@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>