fcoemon: Activate physical port if vlan reply has 0 vlan id
Some switch implementations (eg., HP virtual connect FlexFabric) can send a
vlan notification with vid 0, which is an indication that the switch will send
the packets on the physical interface itself, without any tag. Also, it is
expected that the priority information is delivered directly via outer vlan
tag.  Create fcoe on the physical interface when vlan notification has vid 0.
This change should help the CNAs that support this special switch. This should
not affect the functionality of software stack, as vlan notification with vid 0
is not expected to be received from the regular FCFs.

Signed-off-by: Nithin Nayak Sujir <nsujir@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>