fcoedump: Allow check for 'boot.*' init scripts
SLES is using boot.lldpad and boot.fcoe for their
service scripts. This script has fcoedump.sh check
for the 'boot.*' service scripts as well as for
the 'lldpad' and 'fcoe' non-boot-prefixed service
scripts.

Note that this patch intentionally allows both to
be printed if both scripts exist. This would indicate
to someone debugging that there are duplicate
scripts and there is a problem.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>