fcoemon: Add signal handler to dump fcoemon data structures
This patch modifies the fcoemon signal handler so that it
can call a different routine for new signals. It also has
fcoemon register for the SIGUSR1 signal and when it catches
that signal it cruedly dumps the state of fcoemon's fcoe_port
and fcm_netif lists.

No documentation is provided with this feature as it is a
developer only feature.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>