fcping: Fix misleading --help output
The '--help' option shows that a target is optional
even though it is required. The user _must_ specify,
'-F <FC-ID>' or '-P <WWPN>', or '-N <WWNN>'. The man
page correctly expresses how 'fcping' behaves. This
patch makes the '--help' output match the man page
and more importantly, the command's actual behavior.

This patch also moves the target arguments onto the
same line as the other options. It also changes the
spacing on the first few options from '[-fqx]' to
'[ -fqx ]'. All other options have a space after the
initial brace and before the closing brace.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>