fipvlan: Start fcoe on physical interface if vlan reply has 0 vlan id
This is equivalent fix of 12637e4eeb51ed273e337e42c22f6ac6721dfc20 in fipvlan.

Some switch implementations (eg., HP virtual connect FlexFabric) can send a
vlan notification with vid 0, which is an indication that the switch will send
the packets on the physical interface itself, without any tag. Also, it is
expected that the priority information is delivered directly via outer vlan
tag.  Create fcoe on the physical interface when vlan notification has vid 0.
This change should help the CNAs that support this special switch. This should
not affect the functionality of software stack, as vlan notification with vid 0
is not expected to be received from the regular FCFs.

Signed-off-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>