fcoe-utils: Add exported include/linux/types.h
Currently fcoe-utils relies on the user to have this
kernel header file installed in the include path. If
the user has an unscrubbed header file the fcoe-utils
build will generate warnings. This patch contributes
the scrubbed kernel header file to the fcoe-utils
include/linux path to avoid any compilation warnings.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>