libutil: Share more common routines in libutils.a
Share table manipulation and sysfs scanning code
in libutil. It will be needed by libopenfcoe.a
in a later patch.

Note that there was some confusion around the sa_enum_decode
function. This patch ensures that the found string is returned
in the 'buf' argument and not just as the return value from
the function. Callers were not relying on the return value.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>