fcoeadm: Add '-f|--fcfs' option to list FCFs
This patch adds the '-f|--fcfs' option to the
fcoeadm application. These options allow the
user to list the discovered Fibre Channel
Forwarders (FCFs).

Please see the man page update or --help for
usage.

This patch also reorders the options as they
are listed so that 'interface', 'fcfs', 'target'
and 'lun' are grouped together. Previously,
'scan' was between 'interface' and 'target'.
It seemed to make more sense to have 'scan' closer
to 'reset' and not in the middle of the FCoE
instance listing progression.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>