build: Remove dependency on libnl
fcoe-utils really did not need to be dependent on libnl. It was
only including netlink/netlink.h, not actually using the library,
and the only real need for netlink/netlink.h was to get the
including files stdint.h, linux/netlink.h and linux/rtnetlink.h.
So, just include those directly and forget about libnl.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>