fcoeadm: Fix scsi_target_id access
This fixes a few problems with how the function show_target_info
accesses the scsi_target_id. First off, it tries to read it as a
u32 value, but checks for the special value -1. There are two
problems with that, -1 can't be read by sa_sys_read_u32 (it
will return an error) and if it were to return 0xffffffff in
the u32, the comparison with -1 would always be false. A third
problem is that the error return from sa_sys_read_u32 is not
checked. Fix this by adding a function sa_sys_read_int and
use it properly.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Tested-by: Ross Brattain <ross.b.brattain@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>