fcoemon: Create a function template for the libfcoe module paramater interface
The libfcoe module parameter interface for create, destroy, enable and disable
are being deprecated. The first step is to generalize and abstract the legacy
interface by making calls to it through a function template. The next patch in
this series will add a new template implemenation with calls to the new fcoe_sysfs
based interface.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Marcus Dennis <marcusx.e.dennis@intel.com>