fcoemon: Add fcoe_sysfs based kernel interface support
Add a function template implementation for the new fcoe_sysfs
kernel interfaces. When the /sys/bus/fcoe/ctlr_create file exists
and is writable by fcoemon the new template will be used.

To write to the new sysfs locations the new helper routine
fcm_write_str_to_ctlr_attr is added. Also, some reshuffling
of the sysfs path defines was required.

Also note that that the fcoe_ctlr lookup only needs to happen
if fcoemon is to use the new fcoe_sysfs interfaces.

Signed-off-by: Robert Love <robert.w.love@intel.com>
Tested-by: Marcus Dennis <marcusx.e.dennis@intel.com>