fcoemon: Add attributes to allow format-string parameter checking
Add format attributes to the static function declarations
of the functions print_result and print_err so that their
parameters can be checked against the given format strings.
Also add format attributes to the prototypes for sa_log,
sa_log_debug and sa_log_err. The format attribute on the
prototype for assert_failed was changed to use __format__
and __printf__ to avoid any possible name clash with a
symbol definition in the future.

Signed-off-by: Mark Rustad <mark.d.rustad@intel.com>
Tested-By: Marcus Dennis <marcusx.e.dennis@intel.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>