initd.fedora: Do not start fcoemon twice
After the fix in commit-id cf50e10170b301542c968da6cdc278fd41730820 PID
file no longer exists in /var/run. As a result on initd.fedora fails to
check if fcoemon was already started.  This patch fixes the issue by
using -p option to obtain the PID, and not relying on the PID_FILE to
check if fcoemon is already running

Signed-off-by: Bhanu Prakash Gollapudi <bprakash@broadcom.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>