fipvlan fails on powerpc
fipvlan fails to execute properly on powerpc architecture.

# fipvlan -a
Try 'fipvlan --help' for more information
#

# fipvlan --create --start eth1
Try 'fipvlan --help' for more information
#

In power pc systems, when there are no more arguments to be parsed, getopt_long
should return -1. But gdb output shows that it c is 255 whereas in intel system
it shows -1 properly.

"char" must not be trusted when signing is important!  Different arches (and even
different gcc versions on the same arch) may default char to signed or unsigned.

Ideally, a portable program should always use signed char or unsigned char when
it depends on the signedness of an object. But many programs have been written
to use plain char and expect it to be signed, or expect it to be unsigned,
depending on the machines they were written for. In this case the variable c
is expected to behave as signed and hence using signed would be the safe option.

http://gcc.gnu.org/onlinedocs/gcc-4.8.1/gcc/C-Dialect-Options.html

Signed-off-by: Raju Chakraborty<rajuchak@linux.vnet.ibm.com>
Signed-off-by: Robert Love <robert.w.love@intel.com>