fcoe-utils: automake fixes for dist/distcheck
I thought I'd give make distcheck a try on fcoe-utils and see if it
produced a working release.  It didn't, so here are some fixes for that.

Missing header files added to git are included in the release archive.

The systemd unit files are marked dist_DATA so they get included,
otherwise data files are assumed to be generated at build time and left
out of the release.

A few other missing doc files are added to dist_noinst_DATA.

The manual rules for handling bash completion files are replaced with
dist_DATA definitions letting automake handle generating rules.

Any change of applying build system fixes, and getting a new tag and
release tarball anytime in the near future?

Signed-off-by: Chris Leech <cleech@redhat.com>
Signed-off-by: Vasu Dev <vasu.dev@intel.com>