rtnetlink: Create rtnl socket without multicast grp for sanmac discovery
fcoemon tries to figure out sanmac, of an interface, before trying to start
VLAN discovery. This is done using a rtnl socket (with nl_groups set to
RTNLGRP_LINK) and RTM_GETDCB query to the kernel. Since nl_group is set to
RTNLGRP_LINK the receive handler also receives RTN_NEW_LINK multicast
notifcations and RTN_NEW_LINK notifications are wrongly considered as valid
repose to query. Since no valid sanmac is discovered, fcoemon fails to start
VLAN discovery on this interface.

This is evident when running fcoemon on XenServer, as the interfaces resets
when they are added to the bridge.

Ideally sanmac discovery doesn't require rtnl socket to be opened with
RTNLGRP_LINK and this would avoid receiving multicast notifications.

Signed-off-by: Anoob Soman <anoob.soman@citrix.com>
Reviewed-by: Johannes Thumshirn <jthumshirn@suse.de>
Signed-off-by: Vasu Dev <vasu.dev@intel.com>