fcoemon: Add retry mechanism during fip vlan discovery if socket fails to open.
If we fail to open the FIP socket during FIP VLAN request initiation it may be
due to a transient condition such as the link not being completely ready.  In
this case opening the socket will succeed on a retry so add socket open retry
to the FIP VLAN timeout handler if the socket has not been open yet.

Signed-off-by: Chad Dupius <chad.dupuis@cavium.com>
Signed-off-by: Johannes Thumshirn <jthumshirn@suse.de>